admin panel complete

+++ TextileShop/authentication/views.py
@@ -1,3 +1,6 @@
+
+############################ pasan start###############################
+
 from contextlib import redirect_stderr
 from http.client import HTTPResponse
 from pickle import NONE
@@ -17,6 +20,9 @@ from reportlab.pdfgen import canvas
 from reportlab.lib.units import inch
 from reportlab.lib.pagesizes import letter
 from django.core.mail import send_mail
+from datetime import datetime
+
+import re
 
 
 
@@ -45,6 +51,9 @@ def sendMail(fname,email,empID,password):
     )
 
 
+
+
+##register employee to the system
 def register(request):
 
     if request.method == "POST":
@@ -57,27 +66,39 @@ def register(request):
         password = request.POST.get('password')
         passwordc = request.POST.get('passwordc')
 
-       # employees = 
-        
-        if password != passwordc :
+        ##null point validation
+        if len(empID) == 0  or len(fname) == 0 or len(lname) == 0 or len(email) == 0 or len(position) == 0 or len(password) == 0 or len(passwordc) == 0:
+            messages.success(request,"Please fill all the fields !!! ")
+       
+        else:
+            patt = r'[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,3}'
+            pattern = re.compile(patt)
+            
+            ##validate email
+            if re.match(pattern,email) is None :
+                messages.success(request,"invalid email , try again !!! ")
+                
+            #check password are match
+            elif password != passwordc :
                 messages.success(request,"Password mismatch , try again !!! ")
 
-        elif EmployeesReg.objects.filter(empid = empID).exists() :
+            ##check employee id already in database
+            elif EmployeesReg.objects.filter(empid = empID).exists() :
                 messages.success(request,"Employee ID already exists, try new ID  !!! ")
         
         ##elif EmployeesReg.objects.filter(email = email).exists() :
                # messages.success(request,"Employee email already exists, try new email  !!! ")
 
-        else:
+            else:
                 saveRecord = EmployeesReg()
                 saveRecord.empid = empID
                 saveRecord.fname = fname
                 saveRecord.lname = lname
                 saveRecord.email = email
                 saveRecord.position = position
-                saveRecord.password = make_password(password)
+                saveRecord.password = make_password(password) ##password hashing
                 
-                saveRecord.save()
+                saveRecord.save()  ##save data in DB
 
                 sendMail(fname,email,empID,password)#call mail function
                 messages.success(request,"Employee Registraion sucessfully")
@@ -90,7 +111,8 @@ def register(request):
 
 
 def index(request):
-    return  render(request,"index.html")
+    context = {}
+    return  render(request,"index.html",context)
 
 
 
@@ -102,24 +124,18 @@ def login(request):
         empID = request.POST['empid']
         password = request.POST.get('password')
 
-        employees = EmployeesReg.objects.all()
-
-        
+        employees = EmployeesReg.objects.all()  ##get all the employee details in DB
 
         for emp in employees:   
-            flag = check_password(password,emp.password)
+            flag = check_password(password,emp.password) ##check hashed password
             if emp.empid == empID and flag :
-
-
                 empid = request.POST.get('empid',None)
                 context['empid'] = empid
                 context['fname'] = emp.fname
                 employee['fname'] = emp.fname
                 leave['fname'] = emp.fname
-                
-
-                
-                if emp.position == "admin" :
+                           
+                if emp.position == "admin" :  ## user position admin redire to admin page
                     messages.success(request,"Admin login sucessfully")
                     return  redirect("adminpage")
                 else :
@@ -166,21 +182,40 @@ def adminpage(request):
 
     return render(request,"admin.html",employee)
 
+
+
+
+
+
 def userpage(request):
     return render(request,"user.html",context)
 
     
 
+
+
+
+
+
 def changepassword(request):
+
+    context['navbar'] = "changePassword"
     if request.method == "POST":
         empID = request.POST.get('empid')
         password = request.POST.get('password')
         passwordc = request.POST.get('passwordc')
 
-        if password != passwordc:
-            messages.success(request,"Password mismatch , try again !!! ")
+
+        if len(password) == 0 or len(password) == 0:  ##check passwords are null or not
+            messages.success(request,"Please fill all the fields !!! ")
             return  redirect("changepassword")
+        
+        ##check password are match
         else:
+            if password != passwordc:
+                messages.success(request,"Password mismatch , try again !!! ")
+                return  redirect("changepassword")
+            else:
                 result = EmployeesReg.objects.filter(empid=empID)   
                 for re in result:
                     saveRecord = EmployeesReg()
@@ -206,32 +241,43 @@ def applyleave(request,id):
  
     leave_type = leave_types.objects.all()
     count = Leave.objects.filter(empid = id).count()
-
+    leaves = Leave.objects.filter(empid = id)
     context['leaveType'] = leave_type
     context['count'] = count
-    
+    context['navbar'] = "leavePage"
+    context['leaves'] = leaves
 
+    today = datetime.now()
+    
     if request.method == "POST":
         empID = request.POST.get('empid')
         date = request.POST.get('date')
         reason = request.POST.get('reason')
         leavetype = request.POST.get('leavetype')
 
-        saveRecord = Leave()
-        saveRecord.empid = empID
-        saveRecord.date = date
-        saveRecord.reason = reason
-        saveRecord.leaveType = leavetype
-        saveRecord.status = "pending"
-        saveRecord.save()
-        messages.success(request,"Apply leave sucessfully")
-        return  redirect("userpage")
+        past = datetime.strptime(date, '%Y-%m-%d')
+       
+
+        if len(reason) == 0 or len(date) == 0 :
+            messages.success(request,"Please fill all the feilds")
+        elif past.date() < today.date():
+             messages.success(request,"Please select  valid date ")
+        else:
+            saveRecord = Leave()
+            saveRecord.empid = empID
+            saveRecord.date = date
+            saveRecord.reason = reason
+            saveRecord.leaveType = leavetype
+            saveRecord.status = "pending"
+            saveRecord.save()
+            messages.success(request,"Apply leave sucessfully")
+            return  redirect("userpage")
 
 
     return render(request,"applyleave.html",context)
 
 
-def generatepdf(request):
+#def generatepdf(request):
 
     buf = io.BytesIO()
     c = canvas.Canvas(buf,pagesize=letter,bottomup=0)
@@ -269,7 +315,7 @@ def printFile(request):
 
     employee = EmployeesReg.objects.all()
 
-    template_path = 'pdf.html'
+    template_path = 'pdf.html'  ##get template
     context['pdf'] = employee
 
     response = HttpResponse(content_type = 'application/pdf')
@@ -288,8 +334,8 @@ def printFile(request):
 
 
 def update_emp(request,id):
-    positiondetails = employee_positions.objects.all()
-    employee = EmployeesReg.objects.get(id = id)
+    positiondetails = employee_positions.objects.all() ##get all employee position details
+    employee = EmployeesReg.objects.get(id = id) ##get employee details using id
     context['id'] = employee.id
     context['empid'] = employee.empid
     context['fname'] = employee.fname
@@ -297,7 +343,11 @@ def update_emp(request,id):
     context['email'] = employee.email
     context['position'] = employee.position
     context['positiondetails'] = positiondetails
-    return render(request,"updateEmp.html",context)
+    return render(request,"updateEmp.html",context)  ##load all data to update employee detail template
+
+
+
+
 
 def updateuser(request):
     if request.method == "POST":
@@ -308,6 +358,8 @@ def updateuser(request):
         position = request.POST['position']
 
         result = EmployeesReg.objects.filter(empid=empID)   
+
+
         for re in result:
             saveRecord = EmployeesReg()
             saveRecord.id = re.id
@@ -323,17 +375,49 @@ def updateuser(request):
     return  redirect("adminpage")
 
 
+
+
+
+
+
 def delete_emp(request,id):
-    print(id)
     employee = EmployeesReg.objects.get(id = id)
-    empLeave = Leave.objects.filter(empid = employee.empid)
-    empLeave.delete()
-    employee.delete()
-    messages.success(request,"Delete details sucessfully")
-    return  redirect("adminpage")
+    empLeave = Leave.objects.filter(empid = employee.empid)  
+    context['item'] = employee
+
+    if request.method == "POST":
+        empLeave.delete()
+        employee.delete()
+        messages.success(request,"Delete details sucessfully")
+        return  redirect("adminpage")
+ 
+
+    
+    return render(request,'confirmDelete.html',context)
+
+
+
+
+
+def delete_position(request,id):  
+    position = employee_positions.objects.get(id = id)  
+    employee = EmployeesReg.objects.filter(position = position.name)
+    context['item'] = position
+
+    if request.method == "POST":
+        position.delete()
+        employee.delete()
+        messages.success(request,"Delete details sucessfully")
+        return  redirect("adminpage")
+   
+    return render(request,'deletePosition.html',context)
+
+
+
 
 
 def leaves(request):
+
     result = Leave.objects.all()
     pending_list = Leave.objects.filter(status = "pending") 
     count = EmployeesReg.objects.all().count()
@@ -342,9 +426,6 @@ def leaves(request):
     pending = Leave.objects.filter(status = "pending").count()
     reject = Leave.objects.filter(status = "reject").count()
     approve = Leave.objects.filter(status = "approve").count()
-
-
-    
 
     leave['approve'] = approve
     leave['leaveDetails'] = result
@@ -368,7 +449,6 @@ def leaveMail(fname,email,date,status):
     #server.starttls()
     #server.login('jayanandanafachion@gmail.com','ncipterepthpugjl')
     #server.sendmail('jayanandanafashion@gmail.com',email,subject)
-
 
     subject = "Leave Application"
     message = "Hello " + fname + "\n Your Leave request on   " + date + "\n is  " + status
@@ -429,6 +509,8 @@ def addNewLeave(request):
         return  redirect("leaves")
 
 
+
+
 def addNewEmpPosition(request):
      if request.method == "POST":
         positionType = request.POST['positionType']
@@ -442,12 +524,17 @@ def addNewEmpPosition(request):
 
         return  redirect("adminpage")
 
+
+
 def delete_leave(request,id):
-    leave = Leave.objects.get(id = id)
-   
+    leave = Leave.objects.get(id = id)   
     leave.delete()
     messages.success(request,"Delete details sucessfully")
     return  redirect("leaves")
+
+
+
+############################ pasan  end ###############################
